Adjustes attestor/scripts
